Save work
Now the position on a page is done by finding an element which is the
topmost visible element. I used this because just saving page offset
doesn't work anymore if you resize the page(or does it?). My method is
crap, just use offsets and if it doesn't work after resize, invalidate
the marks if resized.