Make keys work on Chromium too
Cross-browser issues seems not to be totally over yet.
I'm using keydown event to listen for keypresses, it seems to the
best when listening for key combinations and if the user has some
non-ascii keyboard layout. Chromium's non-ascii keyCodes are not the same
as keys you press. For instance, when I press 'ä' and am listeninig for
keydown events, Chromium says the keyCode for that is 222. When you
print it out it is 'Þ'. If I want to show what is my understanding what
they are pressing, I don't know what to do. Firefox does the right
thing, it's KeyboardEvent.key property is the same the character 'ä'
when I press that key.

This commit tries to fix this issue by detecting the browser's
features and using numerical values from different KeyboardEvent
properties.